backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
sqlalchemy==2.0.36
aiofiles==24.1.0
anthropic==0.40.0
python-dotenv==1.0.1
faster-whisper==1.1.0
